+++ src/Poll.cpp
@@ -3,7 +3,7 @@
 // Polling setup
 void Manager::setupPollingforServers()
 {
-	for (int i = 0; i < serverList.size(); i++)
+	for (size_t i = 0; i < serverList.size(); i++)
 	{
 		serverList.at(i).print();
 		for (int j = 0; j < serverList.at(i).getNumOfPorts(); j++)
